AD: Loading and displaying Data for Current Weather Fragment. Done!

+++ app/src/main/java/com/example/weatherapp/model/Coord.kt
@@ -4,6 +4,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Coord(
-    val lon: Long,
-    val lat: Long
+    val lon: Double,
+    val lat: Double
 )
